refactor:batch insert를 UnViewedCafeRepository에서 하도록 수정
