+++ server/src/main/java/com/project/yozmcafe/domain/cafe/CafeCustomRepository.java
@@ -1,7 +1,5 @@
 package com.project.yozmcafe.domain.cafe;
 
-import com.project.yozmcafe.domain.member.Member;
-
 import java.util.List;
 
 public interface CafeCustomRepository {
@@ -9,6 +7,4 @@ public interface CafeCustomRepository {
     List<Cafe> findAllBy(final String cafeName, final String menu, final String address);
 
     List<Cafe> findAllBy(final String cafeName, final String address);
-
-    void saveUnViewedCafes(final List<Cafe> cafes, final Member member);
 }
