+++ server/src/main/java/com/project/yozmcafe/domain/cafe/CafeCustomRepositoryImpl.java
@@ -1,16 +1,10 @@
 package com.project.yozmcafe.domain.cafe;
 
-import com.project.yozmcafe.domain.member.Member;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.core.types.dsl.StringPath;
-import jakarta.transaction.Transactional;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
-import org.springframework.jdbc.core.BatchPreparedStatementSetter;
-import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
 import java.util.List;
 
 import static com.project.yozmcafe.domain.cafe.QCafe.cafe;
@@ -22,12 +16,9 @@ import static io.micrometer.common.util.StringUtils.isBlank;
 public class CafeCustomRepositoryImpl extends QuerydslRepositorySupport implements CafeCustomRepository {
 
     private static final double MATCH_THRESHOLD = 0.0;
-    private static final int INSERT_BATCH_SIZE = 110;
-    private final JdbcTemplate jdbcTemplate;
 
-    public CafeCustomRepositoryImpl(final JdbcTemplate jdbcTemplate) {
+    public CafeCustomRepositoryImpl() {
         super(Cafe.class);
-        this.jdbcTemplate = jdbcTemplate;
     }
 
     public List<Cafe> findAllBy(final String cafeNameWord, final String menuWord, final String addressWord) {
@@ -57,32 +48,5 @@ public class CafeCustomRepositoryImpl extends QuerydslRepositorySupport implemen
         return numberTemplate(Double.class, "function('match_against', {0}, {1})",
                 target, formattedSearchWord)
                 .gt(MATCH_THRESHOLD);
-    }
-
-    @Transactional
-    public void saveUnViewedCafes(final List<Cafe> cafes, final Member member) {
-        for (int i = 0; i < cafes.size(); i += INSERT_BATCH_SIZE) {
-            List<Cafe> subItems = cafes.subList(i, Math.min(i + INSERT_BATCH_SIZE, cafes.size()));
-            batchInsert(subItems, member.getId());
-        }
-    }
-
-    private void batchInsert(final List<Cafe> cafes, final String memberId) {
-        final String sql = "INSERT INTO un_viewed_cafe (cafe_id, member_id) VALUES (?, ?)";
-
-        jdbcTemplate.batchUpdate(sql,
-                new BatchPreparedStatementSetter() {
-                    @Override
-                    public void setValues(PreparedStatement pstmt, int i) throws SQLException {
-                        pstmt.setLong(1, cafes.get(i).getId());
-                        pstmt.setString(2, memberId);
-                    }
-
-                    @Override
-                    public int getBatchSize() {
-                        return cafes.size();
-                    }
-                }
-        );
     }
 }
