+++ server/src/main/java/com/project/yozmcafe/service/UnViewedCafeService.java
@@ -3,6 +3,7 @@ package com.project.yozmcafe.service;
 import com.project.yozmcafe.domain.CafeShuffleStrategy;
 import com.project.yozmcafe.domain.cafe.Cafe;
 import com.project.yozmcafe.domain.cafe.CafeRepository;
+import com.project.yozmcafe.domain.cafe.UnViewedCafeRepository;
 import com.project.yozmcafe.domain.member.Member;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -16,11 +17,13 @@ public class UnViewedCafeService {
     public static final int DEFAULT_UNVIEWED_CAFE_SIZE_CONDITION = 20;
 
     private final CafeRepository cafeRepository;
+    private final UnViewedCafeRepository unViewedCafeRepository;
     private final CafeShuffleStrategy shuffleStrategy;
 
-    public UnViewedCafeService(final CafeRepository cafeRepository,
+    public UnViewedCafeService(final CafeRepository cafeRepository, final UnViewedCafeRepository unViewedCafeRepository,
                                final CafeShuffleStrategy shuffleStrategy) {
         this.cafeRepository = cafeRepository;
+        this.unViewedCafeRepository = unViewedCafeRepository;
         this.shuffleStrategy = shuffleStrategy;
     }
 
@@ -28,6 +31,6 @@ public class UnViewedCafeService {
     public void refillUnViewedCafes(final Member member) {
         final List<Cafe> shuffledCafes = shuffleStrategy.shuffle(cafeRepository.findAll());
         final List<Cafe> unViewedCafes = member.filterAlreadyExist(shuffledCafes);
-        cafeRepository.saveUnViewedCafes(unViewedCafes, member);
+        unViewedCafeRepository.saveUnViewedCafes(unViewedCafes, member);
     }
 }
